Added SequencedCollections class to illustrate Java 21's Sequenced Collection features
The new class, SequencedCollections, is added to the existing 'com.loiane' package, to demonstrate the usage and features of new Java 21's Sequenced Collection API. The file provides examples using Sequenced Collections, Sequenced Sets, and Sequenced Maps. In addition to basic operations, it shows use of some advanced methods such as reversed(), addFirst(), addLast(), removeFirst(), and removeLast(). The purpose is to exhibit potential application of these sequences in programming logic with the newest Java 21 API.