Commit message: Add UnnamedClasses class to illustrate basic Java concepts in Java 21
This commit introduces a new file, UnnamedClasses.java, to the 'com.loiane' package. The file contains a basic main function that prints a "Hello" string to standard output. The purpose of this class is to serve as a basic introduction to Java for beginners and to lay foundations for further, more complex examples. No advanced features are utilized in this commit. Future commits will gradually introduce more complex coding scenarios and new Java features to encourage progressive learning.