Add UnnamedVariables class to demonstrate Java 21 unnamed variable feature
This new class, UnnamedVariables.java, demonstrates the usage of unnamed variables provided from Java 21. It includes two different methods of handling variables in a for loop: before and after Java 21. The class also includes a bug note with a link to the JDK issue, which has been resolved as of version 21.0.1. Besides, it provides an example of formatted date parsing and error handling without using exception variable.